Move product options helper under Spree module
