+++ config/solidus_compare.yml
@@ -38,5 +38,5 @@ ignore:
     file: frontend/app/controllers/spree/content_controller.rb
 - path: app/helpers
   diffs:
-  - hash: 3170659934
-    file: templates/app/helpers/product_options_helper.rb
+  - hash: 4159072969
+    file: templates/app/helpers/spree/product_options_helper.rb
